+++ routes/organizer.js
@@ -7,8 +7,6 @@ var json2csv = require('json2csv').parse;
 var moment = require('moment');
 var fs = require('fs');
 
-const fields = ['goal', 'need', 'current_solution', 'problem', 'solution_ideas', 'current_benefits', 'comment', 'email']
-
 router.get('/', function(req, res, next) {
     // this is for just fetching and showing the table I think
     var userVar = null;
@@ -39,6 +37,7 @@ router.post('/', function(req, res, next) {
             id = parseInt(key);
         }
     });
+
     queries.updateCard([id, req.body.var1, req.body.link12, req.body.var2, req.body.link23, req.body.var3], (response) => {
         if (!response){
             console.log("[Alert] Updating card failed.");
