--- views/neworganizer.pug
@@ -1,35 +1,0 @@
-extends includes/layout
-
-block content
-    h2.text-center Your cards:
-    hr
-    .row.text-center.work-pad
-        if isEmpty
-            h3 There are no cards in the database yet.
-        else
-            each card in data
-                .col-md-3
-                    div.cardborder
-                        p
-                            b Goal:
-                            |  #{card.goal}
-                        p
-                            b Need:
-                            |  #{card.need}
-                        p
-                            b Current Solution:
-                            |  #{card.current_solution}
-                        p
-                            b Problem:
-                            |  #{card.problem}
-                        p
-                            b Comments:
-                            |  #{card.comment}
-                        p
-                            b Created By:
-                            |  #{card.email}
-
-            .container
-                br
-                form.text-center(action='', method='post' style="margin-top: 20px")
-                    input(type='submit', name='export', value='export to csv')
